arrow icon / burger menu
On the home page and the about page, added an arrow icon, and modified JS to allow the arrow to dynamically spin. changed styles so that the hamburger menu would fill 50% of the right side of the screen. When the burger is opened, the hidden class from the main container is removed and overflow hidden is added to the body to prevent vertical scrolling of the contents.